fix(jepa_gist): pool_encoder_states must preserve dtype, not force fp32
pool_encoder_states forced .to(float32), so under a bf16 model the pooled
state was fp32 while the predictor's MLP weights were bf16. During training
autocast masked this; the trainer's post-training val AND the eval gate both
run predict_latent outside autocast, so both raised 'mat1 and mat2 must have
the same dtype, got Float and BFloat16'. The crash lost run_summary.json
(the trainer crashed in the final-val section after saving jepa_final.pt).

Fix: preserve the input dtype (bf16 on GPU, fp32 on CPU). The mean over
d_state is numerically safe in bf16. Verified: 9 unit tests pass; bf16
predict_latent now works outside autocast on cuda.

+++ src/subconscious/jepa_gist.py
@@ -100,7 +100,13 @@ def pool_encoder_states(enc_states: list[Tensor]) -> Tensor:
         raise ValueError("pool_encoder_states called with no state tensors")
     per_layer = []
     for st in enc_states:
-        s = st.to(torch.float32)  # [b, d_state, d_model_enc]
+        # Preserve the input dtype (bf16 on GPU, fp32 on CPU). The mean over
+        # d_state is numerically safe in bf16, and preserving dtype keeps the
+        # pooled tensor compatible with the predictor's MLP weights OUTSIDE
+        # autocast (a forced fp32 cast here would hand a bf16 Linear a fp32
+        # input and raise a dtype-mismatch once autocast is off -- which is
+        # exactly the eval path and the trainer's post-training val).
+        s = st
         if s.dim() == 3:
             per_layer.append(s.mean(dim=1))        # [b, d_model_enc]
         elif s.dim() == 2:
